moved the alpha checking and field to LocationUtils
