+++ src/main/java/codes/cookies/mod/data/profile/ProfileStorage.java
@@ -9,6 +9,7 @@ import java.util.Optional;
 import java.util.concurrent.TimeUnit;
 
 import codes.cookies.mod.api.ApiManager;
+import codes.cookies.mod.utils.skyblock.LocationUtils;
 import com.google.gson.JsonObject;
 import codes.cookies.mod.CookiesMod;
 import codes.cookies.mod.data.DataMigrations;
@@ -22,14 +23,6 @@ import codes.cookies.mod.utils.dev.DevUtils;
 import codes.cookies.mod.utils.exceptions.ExceptionHandler;
 import codes.cookies.mod.utils.json.JsonUtils;
 
-import com.mojang.logging.LogUtils;
-
-import net.hypixel.modapi.HypixelModAPI;
-import net.hypixel.modapi.packet.impl.clientbound.ClientboundHelloPacket;
-import net.hypixel.data.region.Environment;
-
-import net.minecraft.client.MinecraftClient;
-
 import org.jetbrains.annotations.NotNull;
 
 /**
@@ -40,8 +33,6 @@ public class ProfileStorage {
 
 	private static final Path PROFILE_DATA_FOLDER = Path.of("config/cookiesmod/profiles");
 	private static ProfileData profileData;
-
-	private static boolean isOnHypixelAlpha = false;
 
 	/**
 	 * Registers the listeners for automatic profile swapping.
@@ -56,16 +47,6 @@ public class ProfileStorage {
 
 		CookiesMod.getExecutorService().scheduleAtFixedRate(ProfileStorage::saveCurrentProfile, 5, 5,
 				TimeUnit.MINUTES);
-		HypixelModAPI.getInstance().createHandler(ClientboundHelloPacket.class, ProfileStorage::onJoinHypixel);
-	}
-
-	private static void onJoinHypixel(ClientboundHelloPacket clientboundHelloPacket) {
-		if(clientboundHelloPacket.getEnvironment() ==  Environment.BETA) {
-			isOnHypixelAlpha = true;
-			CookiesUtils.sendFailedMessage("You are on Hypixel Alpha, profile data will not be saved!");
-		} else {
-			isOnHypixelAlpha = false;
-		}
 	}
 
 
@@ -77,7 +58,7 @@ public class ProfileStorage {
 			return;
 		}
 
-		if (PlayerStorage.getCurrentPlayer().isEmpty() || SkyblockUtils.getLastProfileId().isEmpty() || isOnHypixelAlpha) {
+		if (PlayerStorage.getCurrentPlayer().isEmpty() || SkyblockUtils.getLastProfileId().isEmpty() || LocationUtils.isOnHypixelAlpha()) {
 			return;
 		}
